[Translation] added FileLoader.
